requirements.txt
fastapi
uvicorn
ccxt
pandas
numpy
pandas-ta
google-genai
python-dotenv
requests
python-dateutil
pydantic
structlog
notion-client
schedule